backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
openpyxl>=3.1.5
reportlab>=4.2.2
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
sse-starlette>=2.1.0
websockets>=12.0
